requirements.txt
aiogram==3.20.0.post0
python-dotenv>=1.0.0
requests>=2.32.0
deep-translator>=1.11.4
aiohttp>=3.12.0